requirements.txt
langchain-experimental
langchain-openai
pandas
numpy
python-dotenv
ipykernel
tabulate
openai